refactor: RequestMaker 삭제, generate 메서드 NetworkManager 로 이전
